REFACTOR:moving function to modeler 3d class
