+++ pyaedt/modeler/modeler3d.py
@@ -1505,3 +1505,105 @@ class Modeler3D(Primitives3D):
             return create_region.SetPropValue("Coordinate System", region_cs)
         except (GrpcApiError, SystemExit):
             return False
+
+
+@pyaedt_function_handler()
+def create_geometry_from_csv(self, csv_file, header_line=2, column_mapping=None, geo_type="cylinder", unit="mm"):
+    """Create geometry from csv file containing geometry data.
+
+    Parameters
+    ----------
+    csv_file : str
+        Csv file containing the geometry details.
+        Expected column headers are dependent on the geometry type:
+
+        - if ``geo_type=='cylinder'``, expected fields are ``'xc'`` (base
+        center x coordinate),``'yc'`` (base center y coordinate), ``'zc'``
+        (base center z coordinate), ``'height'``, ``'radius'``,
+        ``'iradius'`` (inner radius for hollow cylinder), ``'plane'``
+        (plane on which the cylinder base lies).
+    header_line : int, optional
+        The line marking the start of data.
+        The default value is ``2``.
+    column_mapping : dict, optional
+        Map of the csv columns names to the accepted column names.
+        The default value is ``None``.
+    geo_type: str, optional
+        Type of geometry that the csv file describe. Accepted value is: ``"cylinder"``.
+        Default is ``"cylinder"``.
+    unit : str , optional
+        Unit of values used to create geometry. Default unit is ``"mm"``.
+
+    Returns
+    -------
+    list of :class:`pyaedt.modeler.cad.object3d.Object3d`
+    """
+    import pandas as pd
+
+    df = pd.read_csv(csv_file, header=header_line)
+    if column_mapping is not None:
+        df = df.rename(columns=column_mapping)
+    if geo_type == "cylinder":
+        return self._create_cylinder(df, unit=unit)
+    else:
+        raise ValueError("Geometry type not supported yet.")
+
+
+@pyaedt_function_handler()
+def _remove_special_characters(self, text):
+    import re
+
+    sanitized_text = re.sub("[^a-zA-Z0-9_ \n]", "_", text)
+    self.logger.info("Renamed " + text + " to " + sanitized_text)
+    return sanitized_text
+
+
+@pyaedt_function_handler()
+def _length_unit_conversion(self, value, input_units):
+    from pyaedt.generic.constants import unit_converter
+
+    converted_value = unit_converter(value, unit_system="Length", input_units=input_units, output_units="mm")
+    return converted_value
+
+
+@pyaedt_function_handler()
+def _create_cylinder(self, df, unit):
+    cylinders = []
+    for i in range(len(df)):
+        try:
+            name = self._remove_special_characters(df["name"][i])
+            radius = self._length_unit_conversion(df["radius"][i], input_units=unit)
+            height = self._length_unit_conversion(df["height"][i], input_units=unit)
+            cs_axis = self._length_unit_conversion(df["plane"][i], input_units=unit)
+            inner_radius = self._length_unit_conversion(df["iradius"][i], input_units=unit)
+            xc = self._length_unit_conversion(df["xc"][i], input_units=unit)
+            yc = self._length_unit_conversion(df["yc"][i], input_units=unit)
+            zc = self._length_unit_conversion(df["zc"][i], input_units=unit)
+            position = (xc, yc, zc)
+        except KeyError:
+            self.logger.info("The column names in the file do not match the expected names")
+            return False
+
+        if inner_radius == 0:
+            try:
+                outer_cylinder = self.modeler.create_cylinder(cs_axis, position, radius, height, name=name)
+            except Exception:  # pragma: no cover
+                self.logger.info("Unable to create cylinder " + name)
+                return False
+        elif radius > inner_radius:
+            try:
+                outer_cylinder = self.modeler.create_cylinder(cs_axis, position, radius, height, name=name)
+                inner_cylinder = self.modeler.create_cylinder(
+                    cs_axis, position, inner_radius, height, name=name + "_inner"
+                )
+                outer_cylinder.subtract([inner_cylinder], keep_originals=False)
+            except Exception:  # pragma: no cover
+                self.logger.info("Unable to create cylinder " + name)
+                return False
+        else:
+            self.logger.info("Inner radius is greater than outer radius")
+            raise ValueError("Inner radius is greater than outer radius")
+        self.logger.info("Created cylinder: " + name)
+        cylinders.append(outer_cylinder)
+
+    return cylinders
